% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constant_node.R
\name{add_constant_node}
\alias{add_constant_node}
\title{Add a constant node}
\usage{
add_constant_node(graph, value)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using
\code{create_graph}.}

\item{value}{the value to use as a constant; can be either \code{numeric},
\code{character}, or \code{logical}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Add a constant value to the graph, which can either be a
number, TRUE/FALSE, or a string.
}

