% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_function_node.R
\name{add_function_node}
\alias{add_function_node}
\title{Add a function node}
\usage{
add_function_node(graph, name)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using
\code{create_graph}.}

\item{name}{the name of the function as a character object.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Add a function to the graph.
}

