% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vector_node.R
\name{add_vector_node}
\alias{add_vector_node}
\title{Add a vector node}
\usage{
add_vector_node(graph, value)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using
\code{create_graph}.}

\item{value}{the vector to be used; can be either composed of \code{numeric},
\code{character}, or \code{logical} values.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Add a vector to the graph, which can either consist of 2 or more
numbers, TRUE/FALSE values, or strings.
}

