% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph_space.R
\name{create_graph_space}
\alias{create_graph_space}
\title{Create graph space}
\usage{
create_graph_space(output = "visNetwork")
}
\arguments{
\item{output}{the rendering engine to use. Currently defaults to
\code{visNetwork}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Initialize a space for the graph.
}

